% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_dark.R
\name{theme_reveal_dark}
\alias{theme_reveal_dark}
\title{Dark theme for ggplot2 that fits with my reveal.js theme}
\usage{
theme_reveal_dark()
}
\description{
Dark theme for ggplot2 that fits with my reveal.js theme
}
